+++ relevanceai/visualise/projector.py
@@ -337,6 +337,9 @@ class Projector(BatchAPIClient, _Base, DocUtils):
                 [i[vector_label][:label_char_length] + "..." for i in docs],
                 docs,
             )
+        
+        if vector_label:
+            point_labels = self.get_field_across_documents(vector_label, docs)
 
         # Dimension reduce vectors
         vectors = np.array(self.get_field_across_documents(vector_field, docs))
@@ -391,6 +394,7 @@ class Projector(BatchAPIClient, _Base, DocUtils):
             marker_size=marker_size,
             cluster=cluster,
             label_char_length=label_char_length,
+            vector_label=point_labels
         )
 
         layout = self._generate_layout(plot_title=plot_title)
@@ -412,8 +416,7 @@ class Projector(BatchAPIClient, _Base, DocUtils):
 
         else:
             fig = go.Figure(data=plot_data, layout=layout)
-            fig.show()
-        return
+            return fig
 
     def _generate_plot_data(
         self,
@@ -429,6 +432,7 @@ class Projector(BatchAPIClient, _Base, DocUtils):
             None,
         ],
         label_char_length: int,
+        vector_label: str
     ):
         """ """
 
@@ -443,6 +447,7 @@ class Projector(BatchAPIClient, _Base, DocUtils):
                         dims=dims,
                         marker_size=marker_size,
                         label_char_length=label_char_length,
+                        vector_label=vector_label
                     )
                 )
 
@@ -455,6 +460,7 @@ class Projector(BatchAPIClient, _Base, DocUtils):
                     dims=dims,
                     marker_size=marker_size,
                     label_char_length=label_char_length,
+                    vector_label=vector_label
                 )
             )
 
@@ -491,10 +497,10 @@ class Projector(BatchAPIClient, _Base, DocUtils):
         return layout
 
     def _generate_plot_info(
-        self, embedding_df, hover_label, dims, marker_size, label_char_length
+        self, embedding_df, hover_label, dims, marker_size, label_char_length,
+        vector_label
     ):
         import plotly.graph_objects as go
-
         custom_data, hovertemplate = self._generate_hover_template(
             df=embedding_df,
             dims=dims,
@@ -506,10 +512,11 @@ class Projector(BatchAPIClient, _Base, DocUtils):
             "x": embedding_df["x"],
             "y": embedding_df["y"],
             "showlegend": False,
-            "mode": "markers",
+            "mode": "markers+text",
             "marker": {"size": marker_size, "symbol": "circle", "opacity": 0.75},
             "customdata": custom_data,
             "hovertemplate": hovertemplate,
+            "text": vector_label
         }
 
         if dims == 2:
@@ -517,7 +524,8 @@ class Projector(BatchAPIClient, _Base, DocUtils):
 
         else:
             scatter_args["z"] = embedding_df["z"]
-            scatter = go.Scatter3d(**scatter_args)
+            scatter = go.Scatter3d(
+                **scatter_args)
 
         return scatter