+++ docsrc/source/changelog.rst
@@ -7,6 +7,7 @@ v0.26.6
 --------
 
 - Fix missing import in plotting since internalising plots
+- Add support for vector labels
 
 v0.26.5
 ---------